requirements.txt
langchain==0.0.350
openai==1.6.1
streamlit==1.28.2
python-dotenv==1.0.0
youtube-transcript-api==0.6.1
faiss-cpu==1.7.4